% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kabsch.R
\name{kabsch_R}
\alias{kabsch_R}
\title{Kabsch Algorithm in R code}
\usage{
kabsch_R(Q, P)
}
\arguments{
\item{Q}{n x d matrix of points to align to to \code{pm}}

\item{P}{n x d matrix of reference points.}
}
\value{
Matrix \code{qm} rotated and translated so that the ith point
is aligned to the ith point of \code{pm} in the least-squares sense.
}
\description{
Aligns two sets of points via rotation and translation after
solving the superpositioning problem in an Eigenvalue Ansatz.
}
\details{
Given two sets of points, with one specified as the reference set,
the other set will be rotated so that the RMSD between the two is minimized.
The format of the matrix is that there should be one row for each of
n observations, and the number of columns, d, specifies the dimensionality
of the points. The point sets must be of equal size and with the same
ordering, i.e. point one of the second matrix is mapped to point one of
the reference matrix, point two of the second matrix is mapped to point two
of the reference matrix, and so on.
}
\references{
\url{https://en.wikipedia.org/wiki/Kabsch_algorithm}

(C) 2019 Jamie Macpherson and Jens Kleinjung
}
