% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_encode.R
\docType{class}
\name{sadata-class}
\alias{sadata-class}
\alias{sadata}
\title{sa_encode.R}
\description{
Encode a protein structure with the M32K25 structural alphabet.
}
\details{
Given a PDB structure, this function will fit C-alpha
coordinates of all consecutive 4-residue fragments with the canonical
fragments of the M32K25 structural alphabet, selecting the optimal
(lowest RMSD) fragment in each case.
sadata: An S4 class for Structural Alphabet (SA) data.
}
\section{Slots}{

\describe{
\item{\code{fragment_letters:}}{Prototype fragment letters forming SA.}

\item{\code{fragment_coordinates:}}{Coordinates of prototype fragments.}
}}

